+++ src/main/java/com/genius/todoffin/security/config/SecurityConfig.java
@@ -4,8 +4,6 @@ package com.genius.todoffin.security.config;
 import com.genius.todoffin.security.filter.JwtAuthenticationFilter;
 import com.genius.todoffin.security.handler.OAuth2SuccessHandler;
 import com.genius.todoffin.security.service.CustomOAuth2UserService;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.Collections;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -15,8 +13,6 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.CorsUtils;
 
 @Configuration
@@ -24,29 +20,17 @@ import org.springframework.web.cors.CorsUtils;
 @RequiredArgsConstructor
 @EnableWebSecurity
 public class SecurityConfig {
-    private static final String ALLOWED_ORIGIN = "http://localhost:5173";
     private static final String PERMIT_URI[] = {"/v3/**", "/swagger-ui/**", "/api/auth/**"};
     private static final String PERMITTED_ROLES[] = {"USER", "ADMIN"};
+    private final CustomCorsConfigurationSource customCorsConfigurationSource;
     private final CustomOAuth2UserService customOAuthService;
     private final OAuth2SuccessHandler successHandler;
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 
-        http.cors(corsCustomizer -> corsCustomizer.configurationSource(
-                                new CorsConfigurationSource() {
-                                    @Override
-                                    public CorsConfiguration getCorsConfiguration(HttpServletRequest request) {
-                                        CorsConfiguration config = new CorsConfiguration();
-                                        config.setAllowedOrigins(Collections.singletonList(ALLOWED_ORIGIN));
-                                        config.setAllowedMethods(Collections.singletonList("*"));
-                                        config.setAllowCredentials(true);
-                                        config.setAllowedHeaders(Collections.singletonList("*"));
-                                        config.setMaxAge(3600L);
-                                        return config;
-                                    }
-                                }
-                        )
+        http.cors(corsCustomizer -> corsCustomizer
+                        .configurationSource(customCorsConfigurationSource)
                 )
                 .csrf().disable()
                 .httpBasic().disable()
